requirements.txt
google-api-python-client==2.108.0
python-dotenv==1.0.0
streamlit==1.29.0 